+++ squalaetp/forms.py
@@ -158,17 +158,3 @@ class CorvetModalForm(BSModalForm):
             self.add_error('xml_data', _('Invalid XML data'))
             data = None
         return data
-
-
-class ExportCorvetForm(forms.Form):
-    PRODUCTS = [('corvet', 'ALL'), ('ecu', 'ECU'), ('bsi', 'BSI'), ('com', 'COM200x')]
-    FORMATS = [('csv', 'CSV'), ('excel', 'EXCEL')]
-
-    formats = forms.ChoiceField(
-        label='Formats', required=False, choices=FORMATS,
-        widget=forms.Select(attrs={'class': 'form-control mx-sm-3 mb-2'})
-    )
-    products = forms.ChoiceField(
-        label='Produit', required=False, choices=PRODUCTS,
-        widget=forms.Select(attrs={'class': 'form-control mx-sm-3 mb-2'}),
-    )
